Extract budget's portable BENC crypto-core into a shared @commons-systems/crypto-core workspace package so both budget and the new office-hours-snapshot producer can reuse the encrypt/decrypt primitives without duplication.
