+++ budget/package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@commons-systems/analyticsutil": "*",
+    "@commons-systems/crypto-core": "*",
     "@commons-systems/components": "*",
     "@commons-systems/ds": "*",
     "@commons-systems/errorutil": "*",
